Split up util.h, enabled cmd validation, added clang format to nix shell, deleted old vs files
