--- UselessShell.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34723.18
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UselessShell", "UselessShell.vcxproj", "{12DB352C-101B-49DB-8A74-D2E8D2FE5504}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Debug|x64.ActiveCfg = Debug|x64
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Debug|x64.Build.0 = Debug|x64
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Debug|x86.ActiveCfg = Debug|Win32
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Debug|x86.Build.0 = Debug|Win32
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Release|x64.ActiveCfg = Release|x64
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Release|x64.Build.0 = Release|x64
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Release|x86.ActiveCfg = Release|Win32
-		{12DB352C-101B-49DB-8A74-D2E8D2FE5504}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CAA093A6-7362-43FC-857A-C902802EB47B}
-	EndGlobalSection
-EndGlobal
