+++ default.nix
@@ -1,7 +1,7 @@
 with import <nixpkgs> {};
 fastStdenv.mkDerivation {
   name = "env";
-  nativeBuildInputs = [ cmake ];
+  nativeBuildInputs = [ cmake clang-tools ];
   buildInputs = [ gcc glibc pkg-config ];
-  shellHook = "code . && cd build";
+  shellHook = "code . && mkdir -p build/linux && cd build/linux";
 }
