+++ include/util/misc.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void us_sleep(unsigned long ms);
